requirements.txt
llama-index>=0.10.0
llama-index-multi-modal-llms-openai>=0.1.0
llama-index-llms-openai>=0.1.0
llama-index-vector-stores-qdrant>=0.1.0
openai>=1.12.0
python-dotenv>=1.0.0
Pillow>=10.0.0
numpy>=1.26.0
transformers>=4.38.0
torch>=2.2.0
PyMuPDF>=1.22.0
qdrant-client>=1.6.0
ftfy>=6.1.0
regex>=2023.0.0
tqdm>=4.66.0 